feat: enhance popup and core state management with targetVaultId
- Added targetVaultId to ExtensionCoreApp and CoreState for improved state management.
- Updated PopupApp to pass targetVaultId to ExtensionCoreApp, enabling better context handling.
- Refactored PopupView to remove unnecessary complexity, consolidating logic within PopupApp.
- Improved type safety by utilizing AuthorizedContext in popup resolvers and middleware.
- Adjusted storage handling to accommodate appSession in context, enhancing session management.
