+++ core/inpage-provider/background/middleware/authorized.ts
@@ -1,16 +1,11 @@
-import { VaultAppSession } from '@core/extension/storage/appSessions'
 import type { BackgroundMethod } from '@core/inpage-provider/background/interface'
 import type { BackgroundResolver } from '@core/inpage-provider/background/resolver'
-import type { BridgeContext } from '@lib/extension/bridge/context'
 
-import { authorizeContext } from '../core/authorization'
+import { authorizeContext, AuthorizedContext } from '../core/authorization'
 
 export const authorized =
   <K extends BackgroundMethod>(
-    resolver: BackgroundResolver<
-      K,
-      BridgeContext & { appSession: VaultAppSession }
-    >
+    resolver: BackgroundResolver<K, AuthorizedContext>
   ): BackgroundResolver<K> =>
   async params =>
     resolver({
